+++ src/main/java/com/mysite/sbb/MainController.java
@@ -1,4 +1,4 @@
-package com.mysite.sbb.controller;
+package com.mysite.sbb;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
