+++ src/main/java/com/mysite/sbb/answer/Answer.java
@@ -1,5 +1,6 @@
-package com.mysite.sbb.entity;
+package com.mysite.sbb.answer;
 
+import com.mysite.sbb.question.Question;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
