+++ src/main/java/com/mysite/sbb/answer/AnswerRepository.java
@@ -1,6 +1,5 @@
-package com.mysite.sbb.repository;
+package com.mysite.sbb.answer;
 
-import com.mysite.sbb.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnswerRepository extends JpaRepository<Answer, Integer> {
