+++ src/main/java/com/mysite/sbb/answer/AnswerService.java
@@ -1,8 +1,6 @@
-package com.mysite.sbb.service;
+package com.mysite.sbb.answer;
 
-import com.mysite.sbb.entity.Answer;
-import com.mysite.sbb.entity.Question;
-import com.mysite.sbb.repository.AnswerRepository;
+import com.mysite.sbb.question.Question;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
