+++ src/main/java/com/mysite/sbb/question/Question.java
@@ -1,5 +1,6 @@
-package com.mysite.sbb.entity;
+package com.mysite.sbb.question;
 
+import com.mysite.sbb.answer.Answer;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
