+++ src/main/java/com/mysite/sbb/question/QuestionController.java
@@ -1,9 +1,7 @@
-package com.mysite.sbb.controller;
+package com.mysite.sbb.question;
 
 import com.mysite.sbb.AnswerForm;
 import com.mysite.sbb.QuestionForm;
-import com.mysite.sbb.entity.Question;
-import com.mysite.sbb.service.QuestionService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
