+++ src/main/java/com/mysite/sbb/question/QuestionRepository.java
@@ -1,6 +1,5 @@
-package com.mysite.sbb.repository;
+package com.mysite.sbb.question;
 
-import com.mysite.sbb.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
